custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>If typeExtension is unresolved when createElement called then local name is a type</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If an element definition with matching localName, namespace, and TYPE is not registered with token's document, set TYPE to localName">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test (function () {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(value) {
        var obj = doc.createElement(value);        
        var name = 'x-1-' + value;
        var proto = Object.create(obj.constructor.prototype);
        var customElement = doc.createElement(value, name);
        assert_equals(customElement.constructor.prototype.constructor.name, 
                obj.constructor.name, 
                'Wrong custom element type before registration of the custom element type');
        
        doc.registerElement(name, {prototype: proto, extends: value});
        assert_equals(customElement.constructor.prototype.constructor.name, name, 
                'Wrong custom element type after registration of the custom element type');        
    });
}, 'If typeExtension presents in createElement call then type is a type extension');


test (function () {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(value) {
        var obj = doc.createElement(value);
        var name = 'x-2-' + value;
        var proto = Object.create(obj.constructor.prototype);
        var customElement = doc.createElementNS(HTML_NAMESPACE, value, name);
        assert_equals(customElement.constructor.prototype.constructor.name, 
                obj.constructor.name, 
                'Wrong custom element type before registration of the custom element type');
        
        doc.registerElement(name, {prototype: proto, extends: value});
        assert_equals(customElement.constructor.prototype.constructor.name, name, 
                'Wrong custom element type after registration of the custom element type');        
    });
}, 'If typeExtension presents in createElementNS call then type is a type extension');

</script>
</body>
</html>
